remove files and rename from llama
